requirements.txt
click
PyYAML
tqdm
glob2
# requests